1.5.0

+++ elements-hive-for-breakdance/extensions/backgrounds/webgl_kinetic_typography/actions.php
@@ -11,12 +11,12 @@ function addActions( $actions ) {
 				'script' => <<<JS
 					( function() {
 						if ('{{design.elements_hive.backgrounds.background_type}}' == 'webgl_kinetic_typography' ) {
-							const containerEl = document.querySelector('%%SELECTOR%%')
-							const settings = {{design.elements_hive.backgrounds.webgl_kinetic_typography|json_encode}} ?? {}
-							let instance = null
+							const containerEl = document.querySelector('%%SELECTOR%%');
+							const settings = {{design.elements_hive.backgrounds.webgl_kinetic_typography|json_encode}} ?? {};
+							let instance = null;
 							if ( window.EhInstancesManager.instanceExists('EhWebglKineticTypography', '%%ID%%') ) {
-								instance = window.EhInstancesManager.getInstance('EhWebglKineticTypography', '%%ID%%')
-								instance.update(settings)
+								instance = window.EhInstancesManager.getInstance('EhWebglKineticTypography', '%%ID%%');
+								instance.update(settings);
 							} else {
 								const options = {
 									containerEl: containerEl,
@@ -26,17 +26,73 @@ function addActions( $actions ) {
 									placeholderImage: "{{getElementsHivePluginUrl()}}assets/images/placeholders/elements-hive-text.jpg",
 									...settings
 								}
-								const instance = new EhWebglKineticTypography(options)
-								window.EhInstancesManager.registerInstance("EhWebglKineticTypography", "%%ID%%", instance)
+								const instance = new EhWebglKineticTypography(options);
+								window.EhInstancesManager.registerInstance("EhWebglKineticTypography", "%%ID%%", instance);
 							}
 						} else {
 							if ( window?.EhInstancesManager?.instanceExists('EhWebglKineticTypography', '%%ID%%') ) {
-								window.EhInstancesManager.deleteInstance('EhWebglKineticTypography', '%%ID%%')
+								window.EhInstancesManager.deleteInstance('EhWebglKineticTypography', '%%ID%%');
 							}
-							document.querySelector('.eh-webgl-kinetic-typography-%%ID%%')?.remove()
+							document.querySelector('.eh-webgl-kinetic-typography-%%ID%%')?.remove();
 						}
 					}());
 				JS,
+				'dependencies' => [
+					'design.elements_hive.backgrounds'
+				]
+			],
+		],
+	];
+	$actions[] = [
+		'onPropertyChange' => [
+			[
+				'script' => <<<JS
+					( function() {
+						if ('{{design.elements_hive.backgrounds.background_type}}' == 'webgl_kinetic_typography' ) {
+							const settings = {{design.elements_hive.backgrounds.webgl_kinetic_typography|json_encode}} ?? {};
+							if (settings?.apply_to_page == false) {
+								return
+							}
+							let doReset = false
+							const containerEl = document.querySelector('%%SELECTOR%%');
+							const fxLayers = containerEl.querySelector('.eh-fx-layers');
+							const webglKineticTypography = containerEl.querySelector('.eh-webgl-kinetic-typography-%%ID%%');
+							const sectionContainer = containerEl.querySelector('.section-container');
+							const sectionChildren = Array.from(containerEl.children);
+							const kineticIndex = sectionChildren.indexOf(webglKineticTypography);
+							const fxLayersIndex = sectionChildren.indexOf(fxLayers);
+							const sectionContainerIndex = sectionChildren.indexOf(sectionContainer);
+							if ( fxLayers ) {
+								if ( kineticIndex < fxLayersIndex ) {
+									doReset = true
+								}
+							} else {
+								if ( kineticIndex > sectionContainerIndex ) {
+									doReset = true
+								}
+							}
+							if ( doReset ) {
+								if ( window?.EhInstancesManager?.instanceExists('EhWebglKineticTypography', '%%ID%%') ) {
+									window.EhInstancesManager.deleteInstance('EhWebglKineticTypography', '%%ID%%');
+								}
+								document.querySelector('.eh-webgl-kinetic-typography-%%ID%%')?.remove();
+								const options = {
+									containerEl: containerEl,
+									sectionContainer: sectionContainer,
+									eventsContainer: settings?.apply_to_page ? window : containerEl,
+									id: %%ID%%,
+									placeholderImage: "{{getElementsHivePluginUrl()}}assets/images/placeholders/elements-hive-text.jpg",
+									...settings
+								}
+								const instance = new EhWebglKineticTypography(options);
+								window.EhInstancesManager.registerInstance("EhWebglKineticTypography", "%%ID%%", instance);
+							}
+						}
+					}());
+				JS,
+				'dependencies' => [
+					'design.fx_layers'
+				]
 			],
 		],
 	];
